backend/requirements.txt
asgiref==3.9.0
awscli==1.41.4
botocore==1.39.4
colorama==0.4.6
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
docutils==0.19
jmespath==1.0.1
MetaTrader5==5.0.5120
numpy==2.3.1
pillow==11.3.0
psycopg2==2.9.10
pyasn1==0.6.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
rsa==4.7.2
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
fastapi==0.104.1
uvicorn==0.24.0
cryptography==45.0.5
pypaystack2
